+++ src/illion/zdd.h
@@ -59,7 +59,7 @@ zdd_t meet(zdd_t f, zdd_t g);
 zdd_t nonsubsets(zdd_t f, zdd_t g);
 zdd_t nonsupersets(zdd_t f, zdd_t g);
 zdd_t choose_random(zdd_t f, std::vector<elem_t>* stack, int* idum);
-zdd_t choose_best(zdd_t f, const std::vector<int>& weights,
+zdd_t choose_best(zdd_t f, const std::vector<double>& weights,
                   std::set<elem_t>* s);
 void dump(zdd_t f, std::ostream& out);
 void dump(zdd_t f, FILE* fp = stdout);
@@ -75,7 +75,7 @@ void _enum(zdd_t f, std::ostream& out, std::vector<elem_t>* stack, bool* first,
            const std::pair<const char*, const char*>& inner_braces);
 void _enum(zdd_t f, FILE* fp, std::vector<elem_t>* stack, bool* first,
            const std::pair<const char*, const char*>& inner_braces);
-void algo_b(zdd_t f, const std::vector<int>& w, std::vector<bool>* x);
+void algo_b(zdd_t f, const std::vector<double>& w, std::vector<bool>* x);
 intx_t algo_c(zdd_t f);
 zdd_t zuniq(elem_t v, zdd_t l, zdd_t h);
 double ran3(int* idum);
